+++ formiodata/components/__init__.py
@@ -1,0 +1,2 @@
+# Copyright Nova Code (http://www.novacode.nl)
+# See LICENSE file for full licensing details.